+++ extractor/src/main/scala/com/digitalasset/extractor/config/ExtractorConfig.scala
@@ -4,7 +4,13 @@
 package com.digitalasset.extractor.config
 
 import java.util.UUID
+import scalaz.OneAnd
+import scalaz.syntax.foldable._
+import scalaz.syntax.functor._
+import scalaz.std.list._
+import scalaz.std.string._
 
+import com.digitalasset.daml.lf.data.Ref.Party
 import com.digitalasset.ledger.api.v1.ledger_offset.LedgerOffset
 import com.digitalasset.ledger.api.tls.TlsConfiguration
 
@@ -20,7 +26,14 @@ final case class ExtractorConfig(
     ledgerPort: Int,
     from: LedgerOffset,
     to: SnapshotEndSetting,
-    party: String,
+    parties: ExtractorConfig.Parties,
     tlsConfig: TlsConfiguration,
     appId: String = s"Extractor-${UUID.randomUUID().toString}"
-)
+) {
+  @SuppressWarnings(Array("org.wartremover.warts.Any")) // huh?
+  def partySpec: String = parties.widen[String] intercalate ","
+}
+
+object ExtractorConfig {
+  type Parties = OneAnd[List, Party]
+}